(CAT-2484): Stop yaml file from being overwritten
This change is simple in that it moves some of the logic to the PuppetLint.  Every file gets its own PuppetLint object with its own list of problems, checks, etc.  Therefore, each PuppetLint should know itself (*.pp?  *.yml?  *.yaml?) and whether it is a valid candidate for fixing.  Keep all validation and fix logic on the PuppetLint instead of in the bin.rb executable.

Signed-off-by: Gavin Didrichsen <gavin.didrichsen@gmail.com>